zyp：商城dao层基本功能（查询上架上品，可购买商品，进货量，购买量，库存量）实现并且通过测试

+++ src/main/java/cn/edu/usst/cs/campusAid/dto/shop/ProductTransaction.java
@@ -15,5 +15,5 @@ public class ProductTransaction {
      * <p>可以为空，表示维持现价</p>
      */
     Float price;
-    float amount;
+    Long amount;
 }
